Add logging framework, basic main function.
Adds a logging package which is a thin wrapper around the built in slog.
This package exposes a way to define a default logger, and package level
convenience functions for interacting with that default logger.

Also did a refactor of the config package to be able to handle what I am
calling 'dynamic' properties.  In this commit the only one is the logs
file path.  This is dynamic because it should be determined fresh at
every application start, unless specified in the config file.  It also
shouldn't be saved as a 'default' to the config file either.  The log
file path is based on OS-specific expectations.

Finally, a minimal main.go was created to actually be able to e2e
test all of this.