Add episode search capability.
Handles uncommon cases like when AniList considers a season to be one large set
of episodes, while allanime considers them to be two halves and gives each their
only set of 1-n episode numbers.

Currently this lookup is triggered only by ctrl-p on the selected anime,
and the output is only logs.