Add playback starting and monitoring for MPV
There's some serious rough edges here, but it's largely working for now
so I wanted to get a commit in.

Starts playback in MPV, connecting to the JSON IPC socket.  Waits
for events indicating that playback has started and then closes the
'loading' screen to go back to the list view.