Rely solely on the modelStack
When a new model needs to go on the stack, it is created
fresh.  When one is popped, its state is lost.
This helps transition to a more stateless approach.  There is still
some state that needs to be managed at the AppModel level, but the
actual workload models shouldn't need any.